Separate multiplayer classes
