+++ AnoMech/Multiplayer/MultiplayerManager.cs
@@ -45,7 +45,7 @@ namespace AnoMech.Multiplayer;
 // there because Plugin drives it from OnFrameworkUpdate, and every handler
 // reached from RelayClient.MessageReceived (a background receive thread) is
 // marshalled onto it via Plugin.Framework.Run before touching any game state.
-public sealed class MultiplayerManager : IDisposable
+public sealed partial class MultiplayerManager : IDisposable
 {
     private const string CodeAlphabet = "ABCDEFGHJKLMNPQRSTUVWXYZ23456789"; // no 0/O/1/I
     // Mirrors UmadP5ExaflaresScenario.FrameGapCapSeconds -- a peer's P5 debug-bot
@@ -900,1414 +900,4 @@ public sealed class MultiplayerManager : IDisposable
         running = true;
     }
 
-    // ---- Per-frame tick (framework thread; see Plugin.OnFrameworkUpdate) ----
-
-    // Host-only local bookkeeping for "my own run just ended," shared by the
-    // normal in-Tick detection (ActiveScenario went null via Reset/Leave/a
-    // natural finish, still connected -- broadcasts EndMessage right after
-    // this) and Tick()'s own-disconnection-timeout branch above (can't
-    // broadcast anything, every peer has already reached the same conclusion
-    // independently). Does NOT touch the relay/session/roster -- only that
-    // the host's own local run bookkeeping is consistent again.
-    private void EndHostRunLocally()
-    {
-        running = false;
-        // Mirrors StopDebugBotReplay's peer-side reset -- without this, a
-        // debug-bot host would stay bot-driven (or, on a fresh non-multiplayer
-        // solo run afterward, only by luck not still be) past the fight ending.
-        DebugBotControl.Enabled = false;
-        Plugin.GameInstance.PartyMemberKilled -= OnPartyMemberKilledHost;
-        // Without this, Session.Started stays true forever once a run ends --
-        // the Start button in MultiplayerWindow is gated on !Started, so the
-        // lobby would be permanently stuck past its first run with no way to
-        // retry short of leaving and re-hosting a brand new session/code.
-        Session.Started = false;
-    }
-
-    // Host-only: finalize the local run and tell every peer it ended. Shared by
-    // Tick()'s own ActiveScenario-null edge detection and the explicit callers
-    // below -- factored out because that edge only fires ONCE per run ending
-    // (hostScenarioStarted), so a Reset followed by a separate Leave click
-    // previously left the second transition (still-in-zone -> left-the-zone)
-    // completely unobserved: nothing ever told peers the host had actually left,
-    // stranding them in-instance with a Leave button whose LeaveRequestMessage
-    // the host would then also silently drop (see the handler below, gated on
-    // IsInInstance which was by then already false).
-    // returnedToInn is a caller-supplied fact, not something re-derived from
-    // World.Map.IsInInstance here -- Game.Leave()/Reset() both do their actual
-    // work inside a Plugin.Framework.Run(...) callback, which runs on a later
-    // frame, not synchronously. A caller invoking this method right after
-    // Leave() (NotifyLeftInstance, the LeaveRequestMessage handler) would
-    // observe IsInInstance still true -- the pre-Leave state -- and wrongly
-    // broadcast ReturnedToInn=False, telling every peer to Reset() instead of
-    // Leave(). This exact race was why guests stayed stuck in-instance after
-    // the host clicked Leave: the host's own Leave() eventually ran a few
-    // frames later, but the broadcast describing it had already gone out
-    // tagged as a Reset. The Tick() edge-trigger call site below is the one
-    // exception where reading IsInInstance is safe -- it only fires once
-    // ActiveScenario has already gone null on some later tick, by which point
-    // any deferred Leave()/Reset() from earlier has genuinely finished.
-    private void BroadcastRunEnded(bool returnedToInn)
-    {
-        EndHostRunLocally();
-        _ = relay?.SendAsync(Session.ToMessage());
-        DiagnosticLog.Info($"[Multiplayer] Run ended (ReturnedToInn={returnedToInn}) -- broadcasting EndMessage.");
-        _ = relay?.SendAsync(new EndMessage(ReturnedToInn: returnedToInn));
-        LobbyChanged?.Invoke();
-    }
-
-    // Called from the host's own "Leave" button (MainWindow) right after
-    // Plugin.GameInstance.Leave() -- see BroadcastRunEnded's doc comment for why
-    // the Tick()-driven edge trigger can't be relied on here: if the host
-    // Reset() first, that edge already fired and consumed itself, so this
-    // explicit call is the only thing that will ever tell peers about the
-    // Leave() that follows it.
-    public void NotifyLeftInstance()
-    {
-        if (!IsHost || relay is not { IsConnected: true }) return;
-        BroadcastRunEnded(returnedToInn: true);
-    }
-
-    // Subscribed lazily on first Tick rather than in the field initializer that
-    // constructs this class (Plugin.GameInstance isn't set yet at that point) --
-    // one-shot, since Plugin.GameInstance.World (and so World.Map) is a single
-    // long-lived instance for the plugin's lifetime, never recreated per run.
-    // Handlers themselves gate on IsHost so this is a harmless no-op for a peer
-    // or during solo play; see MapEffectMessage/MapDirectorUpdateMessage.
-    private bool mapEventsSubscribed;
-
-    private void SubscribeMapEventsOnce()
-    {
-        if (mapEventsSubscribed) return;
-        mapEventsSubscribed = true;
-        Plugin.GameInstance.World.Map.EffectApplied += (packetFlags, index) =>
-        {
-            if (!IsHost || relay is not { IsConnected: true }) return;
-            DiagnosticLog.Info($"[Multiplayer] Host: broadcasting MapEffect packetFlags=0x{packetFlags:X8} index=0x{index:X}.");
-            _ = relay.SendAsync(new MapEffectMessage(packetFlags, index));
-        };
-        Plugin.GameInstance.World.Map.DirectorUpdated += (category, arg1, arg2, arg3, arg4, arg5, arg6) =>
-        {
-            if (!IsHost || relay is not { IsConnected: true }) return;
-            DiagnosticLog.Info($"[Multiplayer] Host: broadcasting MapDirectorUpdate category=0x{category:X8}.");
-            _ = relay.SendAsync(new MapDirectorUpdateMessage(category, arg1, arg2, arg3, arg4, arg5, arg6));
-        };
-        Plugin.GameInstance.World.Map.WeatherChanged += (weatherId, transition) =>
-        {
-            if (!IsHost || relay is not { IsConnected: true }) return;
-            DiagnosticLog.Info($"[Multiplayer] Host: broadcasting SetWeather weatherId={weatherId} transition={transition}.");
-            _ = relay.SendAsync(new SetWeatherMessage(weatherId, transition));
-        };
-    }
-
-    public void Tick(float deltaSeconds)
-    {
-        SubscribeMapEventsOnce();
-        DrainPendingMessages();
-
-        // Checked before the IsConnected early-return below (this is exactly the
-        // case where relay is down) -- see disconnectedSinceMs's own doc comment.
-        // One-shot: nulled out immediately so this doesn't re-fire (and re-call
-        // Leave()) every subsequent tick while still disconnected/reconnecting.
-        if (IsHost && disconnectedSinceMs is { } since && running && Plugin.GameInstance.World.Map.IsInInstance
-            && Environment.TickCount64 - since > PeerStaleTimeoutMs)
-        {
-            DiagnosticLog.Warn($"[Multiplayer] Disconnected from the relay for over {PeerStaleTimeoutMs / 1000}s while running -- leaving the zone myself (every peer has likely already given up waiting and left on their own).");
-            Plugin.GameInstance.Leave();
-            EndHostRunLocally();
-            disconnectedSinceMs = null;
-            LobbyChanged?.Invoke();
-        }
-
-        if (relay is not { IsConnected: true }) return;
-
-        // Connection-quality tracking runs continuously -- lobby and mid-fight
-        // alike -- so the roster's status indicators are already live before
-        // anyone clicks Start, not just once running is true below.
-        if (IsHost)
-        {
-            pingTimer += deltaSeconds;
-            if (pingTimer >= PingIntervalSeconds)
-            {
-                pingTimer = 0f;
-                SendPingAndRefreshStatuses();
-            }
-
-            if (pendingStartResponses is { Count: > 0 } pending)
-            {
-                startCheckTimer += deltaSeconds;
-                if (startCheckTimer >= StartCheckTimeoutSeconds)
-                {
-                    DiagnosticLog.Info($"[Multiplayer] StartCheck timed out waiting on: {string.Join(", ", pending.Select(Session.NameOf))}.");
-                    foreach (var peerId in pending)
-                        startCheckFailures[peerId] = "no response";
-                    FinishStartCheck();
-                }
-            }
-        }
-        else if (IsSessionNotFound)
-        {
-            // No host-broadcast message ever arrived since joining -- the relay's
-            // room for this code is either empty (mistyped/nonexistent session) or
-            // has only non-host peers in it. Fails fast rather than sitting on
-            // NoHostFoundTimeoutMs's much shorter window than PeerStaleTimeoutMs's
-            // grace period below, which is tuned for a host that WAS confirmed
-            // present going silent -- a worse, rarer case that deserves more benefit
-            // of the doubt against transient lag than a bad code does.
-            DiagnosticLog.Warn($"[Multiplayer] No host responded within {NoHostFoundTimeoutMs / 1000}s of joining session {SessionCode} -- session not found.");
-            LeaveSession();
-            SessionEndReason = "Session not found.";
-            LobbyChanged?.Invoke();
-            return;
-        }
-        else if (IsHostStale)
-        {
-            // A clean "Leave session" click reaches peers via SessionEndedMessage
-            // well within PeerStaleTimeoutMs, so this only fires for a host that
-            // vanished without warning -- a crash, alt-F4, or hard network drop,
-            // where no goodbye message was ever possible. Same end state either
-            // way: nobody's left to resume the fight with, so leave the zone (if
-            // mid-fight) and the session both, rather than sitting on a frozen
-            // roster or a stale zone forever. IsInInstance guard: Leave() ->
-            // Unload() assumes a zone was actually entered (it restores the
-            // real character to the position ZoneSession.Enter() saved) --
-            // running can briefly be true before that deferred entry actually
-            // completes, and calling it too early teleports the real character
-            // to garbage coordinates instead.
-            DiagnosticLog.Warn($"[Multiplayer] Lost contact with the host (no message in {SecondsSinceHostMessage:F1}s, threshold {PeerStaleTimeoutMs / 1000}s) -- leaving.");
-            if (running && Plugin.GameInstance.World.Map.IsInInstance) Plugin.GameInstance.Leave();
-            LeaveSession();
-            SessionEndReason = "Lost contact with the host.";
-            LobbyChanged?.Invoke();
-            return;
-        }
-
-        if (!running) return;
-
-        if (IsHost)
-        {
-            // Edge-triggered against ActiveScenario becoming non-null (see
-            // hostScenarioStarted) before "it's null" is allowed to mean "the
-            // run ended" -- RunScenarioAsHost's deferred completion hasn't
-            // necessarily set it yet on the first Tick() after Start.
-            if (Plugin.GameInstance.ActiveScenario != null)
-            {
-                hostScenarioStarted = true;
-            }
-            else if (!hostScenarioStarted)
-            {
-                return;
-            }
-            else
-            {
-                // Reset/Leave clears ActiveScenario -- stop broadcasting once the local
-                // run has ended rather than spamming empty snapshots (or, worse, a
-                // later unrelated solo run) to peers who are still connected.
-                // Safe to read IsInInstance here (unlike the explicit call sites) --
-                // this only fires once ActiveScenario has gone null on some later
-                // tick, by which point Reset()/Leave()'s deferred work has finished.
-                BroadcastRunEnded(!Plugin.GameInstance.World.Map.IsInInstance);
-                return;
-            }
-            if (!aiReplayStateSent) TrySendAiReplayState();
-            SampleAndBroadcastSnapshot();
-        }
-        else
-        {
-            if (Plugin.GameInstance.World.Map.IsInInstance)
-            {
-                if (!peerEnteredInstance) DiagnosticLog.Info("[Multiplayer] Peer's deferred zone entry completed -- now sending SelfPose.");
-                peerEnteredInstance = true;
-                // Cheap and idempotent past the first successful call (guarded
-                // internally on debugBotReplayStarted) -- simpler than a second
-                // edge-trigger flag alongside peerEnteredInstance.
-                TryStartDebugBotReplay();
-                // P5 Exaflares only: UmadP5ExaflaresAi schedules its dodges onto
-                // debugShadowStateP5.Timeline (a private EventScheduler), which the
-                // real scenario normally drives every frame from its own Tick (see
-                // UmadP5ExaflaresScenario.Tick) -- but a peer never runs
-                // IScenario.Tick at all (Game.RunScenarioInternal never sets
-                // ActiveScenario for isPeer:true), so nothing else would ever
-                // advance it. Mirrors that method's two calls exactly, using the
-                // same raw (EventTimeScale-independent) deltaSeconds this Tick
-                // already receives from Plugin.OnFrameworkUpdate -- capped the
-                // same way UmadP5ExaflaresScenario.Tick caps its own wall-clock
-                // delta, so a hitch/alt-tab frame skips this tick instead of
-                // dumping every still-queued event out at once.
-                if (debugShadowStateP5 is { } p5Shadow && deltaSeconds > 0f && deltaSeconds <= P5ReplayFrameGapCapSeconds)
-                {
-                    p5Shadow.Timeline.Tick(deltaSeconds);
-                    p5Shadow.SpreadTick?.Invoke(deltaSeconds);
-                }
-            }
-            else if (peerEnteredInstance)
-            {
-                DiagnosticLog.Info("[Multiplayer] Peer's zone was unloaded out from under the run (IsInInstance went false) -- stopping locally.");
-                running = false;
-                StopDebugBotReplay();
-                return;
-            }
-            else
-            {
-                // Zone load queued by RunScenarioAsPeer hasn't run yet -- wait
-                // rather than tearing down a run that hasn't truly started.
-                return;
-            }
-            SendSelfPose();
-        }
-    }
-
-    // ---- Host: sampling the live simulation --------------------------------
-
-    // No throttle: Tick already runs at most once per Framework update, so any
-    // interval below the actual frame time is a no-op, and every peer benefits
-    // from a boss position/facing that's as fresh as the host's own frame rate
-    // allows. Bandwidth/CPU scales with host FPS instead of being capped --
-    // fine for a handful of peers.
-    private void SampleAndBroadcastSnapshot()
-    {
-        var world = Plugin.GameInstance.World;
-
-        // UMAD P2 only (harmless no-op elsewhere -- LastState is only non-null for
-        // the scenario that's actually running). See P2LockonsUpdateMessage's doc
-        // comment: UmadP2ForsakenScenario.ReapplyLockons reassigns state.Lockons
-        // dynamically as towers resolve, host-only, so a peer's replay-start
-        // snapshot of it goes stale the first time that happens. Re-broadcast
-        // whenever it actually changes, keyed by a sorted string (Dictionary has
-        // no value equality) so this doesn't spam a message every tick.
-        if (Plugin.GameInstance.Scenarios[Session.ScenarioIndex] is UmadP2ForsakenScenario { LastState: { } p2State })
-        {
-            var lockonsKey = string.Join(",", p2State.Lockons.OrderBy(kv => kv.Key).Select(kv => $"{kv.Key}:{kv.Value}"));
-            if (hostLastBroadcastP2Lockons != lockonsKey)
-            {
-                hostLastBroadcastP2Lockons = lockonsKey;
-                DiagnosticLog.Info($"[Multiplayer] Host: broadcasting P2 Lockons update -- [{lockonsKey}].");
-                _ = relay!.SendAsync(new P2LockonsUpdateMessage(new Dictionary<PartyRole, uint>(p2State.Lockons)));
-            }
-        }
-
-        var liveEnemies = world.Children.OfType<SimEnemy>().Where(e => e.IsActive).ToList();
-        foreach (var stale in hostEnemyNetIds.Keys.Where(e => !liveEnemies.Contains(e)).ToList())
-        {
-            DiagnosticLog.Debug($"[Multiplayer] Host: enemy NetId {hostEnemyNetIds[stale]} ({stale.BNpcBaseId}) no longer active -- dropping from broadcast.");
-            hostEnemyNetIds.Remove(stale);
-            hostEnemyLastLoggedModelState.Remove(stale);
-            hostEnemyLastLoggedStatuses.Remove(stale);
-            hostEnemyLastLoggedAnimationTimeline.Remove(stale);
-        }
-
-        var enemies = new List<EnemyState>(liveEnemies.Count);
-        foreach (var enemy in liveEnemies)
-        {
-            if (!hostEnemyNetIds.TryGetValue(enemy, out var netId))
-            {
-                netId = nextEnemyNetId++;
-                hostEnemyNetIds[enemy] = netId;
-                DiagnosticLog.Info($"[Multiplayer] Host: broadcasting new enemy NetId {netId} -- BNpcBase {enemy.BNpcBaseId}, pos {enemy.Position}, visible {enemy.Visible}.");
-            }
-            var cfg = enemy.SpawnConfig;
-            var modelState = enemy.ModelState;
-            if (!hostEnemyLastLoggedModelState.TryGetValue(enemy, out var lastLogged) || lastLogged != modelState)
-            {
-                hostEnemyLastLoggedModelState[enemy] = modelState;
-                DiagnosticLog.Info($"[Multiplayer] Host: enemy NetId {netId} (BNpcBase {enemy.BNpcBaseId}) ModelState -> 0x{modelState:X2}.");
-            }
-            var statusSnapshot = enemy.ActiveStatusSnapshot;
-            var statusKey = string.Join(",", statusSnapshot.Select(s => $"{s.StatusId}:{s.Stacks}"));
-            if (!hostEnemyLastLoggedStatuses.TryGetValue(enemy, out var lastStatusKey) || lastStatusKey != statusKey)
-            {
-                hostEnemyLastLoggedStatuses[enemy] = statusKey;
-                DiagnosticLog.Info($"[Multiplayer] Host: enemy NetId {netId} (BNpcBase {enemy.BNpcBaseId}) statuses -> [{statusKey}].");
-            }
-            if (enemy.AnimationTimelineId is { } timelineId
-                && (!hostEnemyLastLoggedAnimationTimeline.TryGetValue(enemy, out var lastTimeline) || lastTimeline != timelineId))
-            {
-                hostEnemyLastLoggedAnimationTimeline[enemy] = timelineId;
-                DiagnosticLog.Info($"[Multiplayer] Host: enemy NetId {netId} (BNpcBase {enemy.BNpcBaseId}) AnimationTimelineId -> 0x{timelineId:X4}.");
-            }
-            var newLockonVfxIds = enemy.DrainPendingLockonVfxIds();
-            if (newLockonVfxIds.Count > 0)
-                DiagnosticLog.Info($"[Multiplayer] Host: enemy NetId {netId} (BNpcBase {enemy.BNpcBaseId}) NewLockonVfxIds -> [{string.Join(",", newLockonVfxIds)}].");
-            var (castTargetEnemyNetId, castTargetRole) = ResolveTargetId(world, enemy.CastTargetId);
-            var (instantTargetEnemyNetId, instantTargetRole) = ResolveTargetId(world, enemy.LastInstantCastTargetId);
-            enemies.Add(new EnemyState(
-                netId, enemy.BNpcBaseId, cfg.NameId, cfg.Level, cfg.Targetable, enemy.EnemyListMode,
-                cfg.ModelCharaId, cfg.Scale, cfg.HitboxRadius, cfg.InitialModeAttributeFlags, enemy.Visible, modelState,
-                statusSnapshot.Select(s => new EnemyStatusState(s.StatusId, s.Stacks, s.RemainingTime)).ToList(),
-                enemy.AnimationTimelineId, newLockonVfxIds,
-                enemy.Position.X, enemy.Position.Y, enemy.Position.Z, enemy.Rotation,
-                enemy.IsCasting, enemy.CastSeq, enemy.CastActionId, enemy.CastTotalSeconds, enemy.CastOmenDelay,
-                enemy.CastTargetLocation?.X, enemy.CastTargetLocation?.Y, enemy.CastTargetLocation?.Z,
-                castTargetEnemyNetId, castTargetRole,
-                enemy.LastInstantCastSeq, enemy.LastInstantCastActionId,
-                enemy.LastInstantCastTargetLocation?.X, enemy.LastInstantCastTargetLocation?.Y, enemy.LastInstantCastTargetLocation?.Z,
-                instantTargetEnemyNetId, instantTargetRole));
-        }
-
-        var liveTethers = world.Children.OfType<SimTether>().Where(t => t.IsActive).ToList();
-        foreach (var stale in hostTetherNetIds.Keys.Where(t => !liveTethers.Contains(t)).ToList())
-        {
-            DiagnosticLog.Debug($"[Multiplayer] Host: tether NetId {hostTetherNetIds[stale]} no longer active -- dropping from broadcast.");
-            hostTetherNetIds.Remove(stale);
-        }
-
-        var tethers = new List<TetherState>(liveTethers.Count);
-        foreach (var tether in liveTethers)
-        {
-            var (aEnemy, aRole) = ResolveEnd(world, tether.A);
-            var (bEnemy, bRole) = ResolveEnd(world, tether.B);
-            if (!hostTetherNetIds.TryGetValue(tether, out var netId))
-            {
-                netId = nextTetherNetId++;
-                hostTetherNetIds[tether] = netId;
-                DiagnosticLog.Info($"[Multiplayer] Host: broadcasting new tether NetId {netId} (TetherId {tether.TetherId}) -- A={(aEnemy is { } ae ? $"enemy#{ae}" : aRole?.ToString() ?? "null")}, B={(bEnemy is { } be ? $"enemy#{be}" : bRole?.ToString() ?? "null")}.");
-            }
-            tethers.Add(new TetherState(netId, tether.TetherId, aEnemy, aRole, bEnemy, bRole));
-        }
-
-        var roles = new List<RoleState>(8);
-        foreach (var role in Enum.GetValues<PartyRole>())
-        {
-            var member = world.Party.Get(role);
-            var dead = member is ISimPartyMember { Dead: true };
-            IReadOnlyList<EnemyStatusState> statuses = [];
-            IReadOnlyList<uint> newLockonVfxIds = [];
-            if (member != null)
-            {
-                var statusSnapshot = member.ActiveStatusSnapshot;
-                var statusKey = string.Join(",", statusSnapshot.Select(s => $"{s.StatusId}:{s.Stacks}"));
-                if (!hostRoleLastLoggedStatuses.TryGetValue(role, out var lastStatusKey) || lastStatusKey != statusKey)
-                {
-                    hostRoleLastLoggedStatuses[role] = statusKey;
-                    DiagnosticLog.Info($"[Multiplayer] Host: role {role} statuses -> [{statusKey}].");
-                }
-                newLockonVfxIds = member.DrainPendingLockonVfxIds();
-                if (newLockonVfxIds.Count > 0)
-                    DiagnosticLog.Info($"[Multiplayer] Host: role {role} NewLockonVfxIds -> [{string.Join(",", newLockonVfxIds)}].");
-                statuses = statusSnapshot.Select(s => new EnemyStatusState(s.StatusId, s.Stacks, s.RemainingTime)).ToList();
-            }
-            else
-            {
-                hostRoleLastLoggedStatuses.Remove(role);
-            }
-            roles.Add(new RoleState(role, member != null, dead,
-                member?.Position.X ?? 0f, member?.Position.Y ?? 0f, member?.Position.Z ?? 0f, member?.Rotation ?? 0f,
-                statuses, newLockonVfxIds));
-        }
-
-        var liveEventObjects = world.Children.OfType<SimEventObject>().Where(o => o.IsActive).ToList();
-        foreach (var stale in hostEventObjectNetIds.Keys.Where(o => !liveEventObjects.Contains(o)).ToList())
-        {
-            DiagnosticLog.Debug($"[Multiplayer] Host: event object NetId {hostEventObjectNetIds[stale]} (EObj 0x{stale.EObjRowId:X}) no longer active -- dropping from broadcast.");
-            hostEventObjectNetIds.Remove(stale);
-        }
-
-        var eventObjects = new List<EventObjectState>(liveEventObjects.Count);
-        foreach (var eo in liveEventObjects)
-        {
-            if (!hostEventObjectNetIds.TryGetValue(eo, out var netId))
-            {
-                netId = nextEventObjectNetId++;
-                hostEventObjectNetIds[eo] = netId;
-                DiagnosticLog.Info($"[Multiplayer] Host: broadcasting new event object NetId {netId} -- EObj 0x{eo.EObjRowId:X}, pos {eo.Position}, state {eo.CurrentState}.");
-            }
-            eventObjects.Add(new EventObjectState(
-                netId, eo.EObjRowId, eo.VisibleState, eo.CurrentState,
-                eo.Position.X, eo.Position.Y, eo.Position.Z, eo.Rotation));
-        }
-
-        _ = relay!.SendAsync(new WorldSnapshotMessage(enemies, tethers, roles, eventObjects));
-    }
-
-    private (int? enemyNetId, PartyRole? role) ResolveEnd(SimWorld world, SimCharacter? c)
-    {
-        if (c is null) return (null, null);
-        if (c is SimEnemy e) return hostEnemyNetIds.TryGetValue(e, out var id) ? (id, null) : (null, null);
-        foreach (var role in Enum.GetValues<PartyRole>())
-            if (ReferenceEquals(world.Party.Get(role), c)) return (null, role);
-        return (null, null);
-    }
-
-    // Same job as ResolveEnd, but for a Cast() target: SimCast only ever stores the raw
-    // GameObjectId it was given (see SimCast.TargetId's doc comment for why that number
-    // means nothing to a peer on its own), so this resolves by ID equality against the
-    // host's own local party/enemy set instead of ResolveEnd's reference equality.
-    private (int? enemyNetId, PartyRole? role) ResolveTargetId(SimWorld world, GameObjectId? targetId)
-    {
-        if (targetId is not { } id) return (null, null);
-        foreach (var role in Enum.GetValues<PartyRole>())
-            if (world.Party.Get(role)?.GameObjectId == id) return (null, role);
-        foreach (var (enemy, netId) in hostEnemyNetIds)
-            if (enemy.GameObjectId == id) return (netId, null);
-        return (null, null);
-    }
-
-    private void OnPartyMemberKilledHost(PartyRole role, string cause)
-        => _ = relay?.SendAsync(new RoleKilledMessage(role, cause));
-
-    // True once a claimed peer hasn't been heard from for PeerStaleTimeoutMs.
-    // Host reads its own ground-truth peerLastSeenMs; a peer reads the same
-    // number as last relayed by the host (PeerStatusMessage) -- either way
-    // this surfaces in MultiplayerWindow so a dropped connection shows as
-    // something other than "their puppet just stopped moving."
-    public bool IsPeerStale(Guid peerId) => IsHost
-        ? peerLastSeenMs.TryGetValue(peerId, out var lastSeen) && Environment.TickCount64 - lastSeen > PeerStaleTimeoutMs
-        : peerStatuses.TryGetValue(peerId, out var entry) && entry.SecondsSinceLastSeen * 1000f > PeerStaleTimeoutMs;
-
-    // Host-only, every PingIntervalSeconds regardless of running: pings every
-    // claimed peer, rebuilds the display-ready status snapshot from whatever
-    // was last measured (a full cycle behind the very latest Pong, which is
-    // fine for a coarse indicator), and broadcasts it so peers' rosters match.
-    private void SendPingAndRefreshStatuses()
-    {
-        var nowMs = Environment.TickCount64;
-        _ = relay!.SendAsync(new PingMessage(nowMs));
-
-        peerStatuses.Clear();
-        foreach (var peerId in Session.ClaimedBy.Values.Distinct())
-        {
-            if (peerId == MyPeerId || !peerLastSeenMs.TryGetValue(peerId, out var lastSeen)) continue;
-            var latency = peerLatencyMs.TryGetValue(peerId, out var ms) ? ms : (float?)null;
-            peerStatuses[peerId] = new PeerStatusEntry(latency, (nowMs - lastSeen) / 1000f);
-        }
-        _ = relay.SendAsync(new PeerStatusMessage(new Dictionary<Guid, PeerStatusEntry>(peerStatuses)));
-
-        CheckPeerLiveness();
-    }
-
-    private void CheckPeerLiveness()
-    {
-        foreach (var (role, peerId) in Session.ClaimedBy)
-        {
-            if (peerId == MyPeerId) continue;
-            var stale = IsPeerStale(peerId);
-            if (stale && warnedStalePeers.Add(peerId))
-            {
-                DiagnosticLog.Warn($"[Multiplayer] {Session.NameOf(peerId)} ({role}) hasn't reported in over {PeerStaleTimeoutMs / 1000}s -- likely disconnected.");
-                // Mid-fight, a silently-vanished party member dooms the run the same
-                // way an explicit "Leave session" click does (see RemovePeer's mid-
-                // fight branch) -- but unlike that click, going stale isn't
-                // necessarily permanent (a network blip, not a deliberate leave), so
-                // only end the run here; their role claim and roster slot are left
-                // alone so a reconnect (their own client already retries the same
-                // relay/session automatically -- see BeginReconnect) drops them back
-                // into a normal lobby instead of finding their role already handed
-                // to someone else. IsInInstance guard: Leave() -> Unload() assumes a
-                // zone was actually entered, same reasoning as everywhere else this
-                // guard appears in this file.
-                if (running && Plugin.GameInstance.World.Map.IsInInstance)
-                {
-                    DiagnosticLog.Info($"[Multiplayer] Ending the run because {Session.NameOf(peerId)} went stale mid-fight.");
-                    Plugin.GameInstance.Leave();
-                }
-            }
-            else if (!stale && warnedStalePeers.Remove(peerId))
-                DiagnosticLog.Info($"[Multiplayer] {Session.NameOf(peerId)} ({role}) is reporting in again.");
-        }
-    }
-
-    // ---- Peer: reporting our own pose --------------------------------------
-
-    // No throttle, same reasoning as SampleAndBroadcastSnapshot: this feeds
-    // the host's belief about where a real player is (mechanic checks, other
-    // peers' view of them), and it's one GUID + 4 floats, cheap enough to send
-    // every Tick regardless of peer count.
-    private void SendSelfPose()
-    {
-        var player = Plugin.GameInstance.World.Party.Player;
-        if (player == null) return;
-        _ = relay!.SendAsync(new SelfPoseMessage(MyPeerId, player.Position.X, player.Position.Y, player.Position.Z, player.Rotation));
-    }
-
-    // ---- Host: applying a peer's reported pose to their puppet -------------
-
-    private void OnSelfPoseReceived(SelfPoseMessage msg)
-    {
-        if (!IsHost) return;
-        if (Session.RoleOf(msg.PeerId) is not { } role)
-        {
-            DiagnosticLog.Debug($"[Multiplayer] SelfPose from {msg.PeerId} but they hold no claimed role -- dropping.");
-            return;
-        }
-        if (Plugin.GameInstance.World.Party.Get(role) is SimNetworkPuppet puppet)
-            puppet.ApplyNetworkPose(new Vector3(msg.X, msg.Y, msg.Z), msg.Rotation);
-        else
-            DiagnosticLog.Debug($"[Multiplayer] SelfPose from {Session.NameOf(msg.PeerId)} ({role}) but that slot isn't a SimNetworkPuppet -- dropping.");
-    }
-
-    // ---- Peer: applying a world snapshot ------------------------------------
-
-    private void OnWorldSnapshotReceived(WorldSnapshotMessage snap)
-    {
-        if (IsHost) return;
-        // Gated on peerEnteredInstance for the same reason TryStartDebugBotReplay
-        // already is: RunScenarioAsPeer's real zone entry (MapController.TryLoad)
-        // is deferred, so a snapshot can arrive and spawn enemies into
-        // Plugin.GameInstance.World before that load has actually happened -- and
-        // the load that follows moments later tears the zone down and rebuilds it,
-        // destroying those doppels out from under peerEnemies without clearing its
-        // bookkeeping. Every NetId is now considered "already spawned" (it's a key
-        // in peerEnemies), so the "first snapshot -- spawning local doppel" branch
-        // below never fires for them again for the rest of the run: the enemies
-        // are gone from the peer's world permanently. Confirmed via an
-        // AnoMech-DamageDebug dump (UMAD P4 Kefka Says, three-person session): the
-        // three bosses spawned once from a snapshot that arrived 16ms before
-        // "TryLoad: freshLoad=True", then the peer's dump over a minute later
-        // showed "Enemies currently in world: (none)" for the entire rest of the
-        // fight. Snapshots arrive continuously (multiple times a second), so
-        // dropping the ones that land in this narrow pre-load window costs nothing
-        // -- the very next one after zone entry completes finds peerEnemies still
-        // empty and spawns everything fresh, into the world that's actually going
-        // to stick around.
-        if (!peerEnteredInstance) return;
-        var world = Plugin.GameInstance.World;
-
-        var seenEnemyIds = new HashSet<int>();
-        foreach (var e in snap.Enemies)
-        {
-            seenEnemyIds.Add(e.NetId);
-            if (!peerEnemies.TryGetValue(e.NetId, out var enemy))
-            {
-                var config = new EnemySpawnConfig(
-                    e.BNpcBaseId, e.NameId, e.Level, e.Targetable, e.EnemyList, e.Visible,
-                    new Placement(new Vector3(e.X, e.Y, e.Z), e.Rotation),
-                    e.ModelCharaId, e.Scale, e.HitboxRadius, e.InitialModeAttributeFlags);
-                DiagnosticLog.Info($"[Multiplayer] Peer: first snapshot of enemy NetId {e.NetId} -- BNpcBase {e.BNpcBaseId}, pos ({e.X:F2},{e.Y:F2},{e.Z:F2}), rot {e.Rotation:F2}, visible {e.Visible} -- spawning local doppel.");
-                enemy = world.SpawnEnemy(config);
-                if (enemy == null)
-                {
-                    DiagnosticLog.Warn($"[Multiplayer] Peer: SpawnEnemy returned null for NetId {e.NetId} (BNpcBase {e.BNpcBaseId}) -- skipping this enemy.");
-                    continue;
-                }
-                peerEnemies[e.NetId] = enemy;
-            }
-            // Smoothed in Tick rather than teleported here -- see SimEnemy.
-            // ApplyNetworkPosition/TickNetworkPosition. A hard SetPosition every
-            // snapshot made boss movement visibly stutter for peers; the first
-            // snapshot's "spawn" branch above already places the doppel
-            // exactly here via EnemySpawnConfig.Placement, so this call is a no-op
-            // distance-wise on that first tick.
-            enemy.ApplyNetworkPosition(new Vector3(e.X, e.Y, e.Z), e.Rotation);
-            enemy.SetVisible(e.Visible);
-            // Re-issued only on an actual change -- SetModelState's native rebuild
-            // briefly disables/re-enables drawing, so calling it every snapshot
-            // even when unchanged would flicker the model. A scenario's
-            // mid-fight SetModelState calls (Kefka's grow transformation, Omega-M's
-            // phase swaps, etc.) are otherwise a purely local Timeline write the
-            // host never has any other reason to tell peers about -- without this a
-            // peer's doppel just stays on whatever model it first spawned with.
-            if (!peerEnemyModelState.TryGetValue(e.NetId, out var lastModelState) || lastModelState != e.ModelState)
-            {
-                peerEnemyModelState[e.NetId] = e.ModelState;
-                DiagnosticLog.Info($"[Multiplayer] Peer: enemy NetId {e.NetId} (BNpcBase {e.BNpcBaseId}) ModelState -> 0x{e.ModelState:X2}.");
-                enemy.SetModelState(e.ModelState);
-            }
-            // Reconciled against the host's set every snapshot -- AddStatus/RemoveStatus
-            // are cheap direct StatusManager writes (no model rebuild involved), unlike
-            // ModelState above, so there's no need to gate the actual application on a
-            // change; only the log line below is edge-triggered. Without this, a
-            // scenario's stack-based statuses (e.g. UMAD P3's "Max" grow status) never
-            // reach a peer's doppel at all -- it stays Position/Visible/ModelState-correct
-            // but visually un-grown.
-            var currentStatuses = enemy.ActiveStatusSnapshot;
-            foreach (var target in e.Statuses)
-            {
-                if (currentStatuses.Any(s => s.StatusId == target.StatusId && s.Stacks == target.Stacks)) continue;
-                enemy.AddStatus(target.StatusId, duration: target.RemainingTime, stacks: target.Stacks, overrideStacks: true);
-            }
-            foreach (var current in currentStatuses)
-            {
-                if (e.Statuses.Any(s => s.StatusId == current.StatusId)) continue;
-                enemy.RemoveStatus(current.StatusId);
-            }
-            var statusKey = string.Join(",", e.Statuses.Select(s => $"{s.StatusId}:{s.Stacks}"));
-            if (!peerEnemyLastLoggedStatuses.TryGetValue(e.NetId, out var lastStatusKey) || lastStatusKey != statusKey)
-            {
-                peerEnemyLastLoggedStatuses[e.NetId] = statusKey;
-                DiagnosticLog.Info($"[Multiplayer] Peer: enemy NetId {e.NetId} (BNpcBase {e.BNpcBaseId}) statuses -> [{statusKey}].");
-            }
-            // Rising-edge trigger: reuses the real SimCast pipeline (cast bar +
-            // omen VFX) rather than faking either, so timing/placement come from
-            // the same code path solo play already exercises. targetLocation is
-            // threaded through for ground-targeted casts (e.g. BlizzardIII spread
-            // markers) -- omitting it would anchor the telegraph at the caster's
-            // own position instead of the intended ground spot (see NativeCast).
-            // castSeconds is threaded through for the same reason: leaving it null
-            // makes SimCast.Start fall back to a Lumina sheet lookup, which for a
-            // scenario's synthetic helper-enemy action IDs either resolves to a
-            // duration that has nothing to do with what the host scripted (the
-            // telegraph runs too short, out of sync with the host's real damage
-            // timing) or isn't in the sheet at all (Start() logs a warning and the
-            // cast never begins on the peer -- the animation silently never plays).
-            // omenDelay is threaded through too -- left at its 0f default, a cast
-            // like Damning Edict (scripted with omenDelay: 4.1f so its telegraph
-            // only shows for the last ~0.9s of a 5s cast) would instead show that
-            // telegraph for the entire cast on a peer's screen.
-            // targetId is resolved via ResolvePeerEnd (same helper tethers already
-            // use) from the host's role/enemy-NetId translation of its own raw
-            // GameObjectId -- see SimCast.TargetId's doc comment for why the ID
-            // itself can't just cross the network. Omitting it entirely (as the
-            // original TargetLocation-only fix did) left entity-targeted casts like
-            // UMAD P3's Thunder III tankbuster with no target at all on a peer, so
-            // NativeActionEffect's NumTargets went to 0 and the hit-react animation
-            // that's supposed to play on the tank being hit never showed up there.
-            // Dedupes off CastSeq actually changing rather than off IsCasting's rising
-            // edge (host casting && my own doppel isn't) -- see EnemyState.CastSeq's
-            // doc comment for why comparing the host's real cast timer against the
-            // peer's own independently-running replayed one let the exact same cast
-            // replay twice. Guarded on seq > 0 for the same first-connect reason as
-            // LastInstantCastSeq below (don't replay the zero-value default).
-            if (e.CastSeq > 0
-                && (!peerEnemyLastCastSeq.TryGetValue(e.NetId, out var lastCastSeq) || lastCastSeq != e.CastSeq))
-            {
-                peerEnemyLastCastSeq[e.NetId] = e.CastSeq;
-                // Cast()'s omen/telegraph placement (and its own log line) both read
-                // Position/Rotation directly, but ApplyNetworkPosition above only
-                // recorded this snapshot's pose as an interpolation target -- the
-                // fields themselves only catch up gradually, once per frame, in
-                // TickNetworkPosition. A boss that repositions/reorients and casts in
-                // the same host tick (e.g. UMAD P3's Black Hole Face()-then-Cast) would
-                // otherwise have its replayed telegraph drawn from wherever the peer's
-                // doppel was still interpolating from -- see the LastInstantCastSeq
-                // branch below for the observed symptom. Snap to the snapshot's
-                // authoritative pose right before replaying so the telegraph always
-                // matches what the host actually cast, at the cost of skipping one
-                // frame of smoothing exactly when a cast fires.
-                enemy.SetPosition(new Placement(new Vector3(e.X, e.Y, e.Z), e.Rotation));
-                var targetLocation = e.CastTargetX is { } tx && e.CastTargetY is { } ty && e.CastTargetZ is { } tz
-                    ? new Vector3(tx, ty, tz)
-                    : (Vector3?)null;
-                var targetId = ResolvePeerEnd(world, e.CastTargetEnemyNetId, e.CastTargetRole)?.GameObjectId;
-                enemy.Cast(e.CastActionId, targetLocation: targetLocation, castSeconds: e.CastSeconds, omenDelay: e.CastOmenDelay, targetId: targetId);
-            }
-            // Edge-triggered on a monotonic counter rather than IsCasting's rising
-            // edge -- an instant cast (e.g. Nothingness) never makes IsCasting go
-            // true at all (see SimCast.LastInstantCastSeq's doc comment), so this is
-            // the only signal a peer has that one happened. Guarded on seq > 0 so a
-            // peer that just connected doesn't replay the zero-value default the
-            // instant it sees its first snapshot for this enemy.
-            if (e.LastInstantCastSeq > 0
-                && (!peerEnemyLastInstantCastSeq.TryGetValue(e.NetId, out var lastInstantSeq) || lastInstantSeq != e.LastInstantCastSeq))
-            {
-                peerEnemyLastInstantCastSeq[e.NetId] = e.LastInstantCastSeq;
-                // Same stale-pose race as the rising-edge branch above, but far more
-                // visible here: this is exactly the path UMAD P3's Black Hole uses
-                // (Face(tether) 0.1s before an instant Cast(Nothingness), a rect 125x6
-                // line AOE) -- a guest's doppel hadn't finished rotating to the new
-                // tether target yet when the cast replayed, so Nothingness fired along
-                // the doppel's old facing instead, observed as a line AOE seemingly
-                // cutting through the arena center that the host never actually cast
-                // that way.
-                enemy.SetPosition(new Placement(new Vector3(e.X, e.Y, e.Z), e.Rotation));
-                var instantTargetLocation = e.LastInstantCastTargetX is { } itx && e.LastInstantCastTargetY is { } ity && e.LastInstantCastTargetZ is { } itz
-                    ? new Vector3(itx, ity, itz)
-                    : (Vector3?)null;
-                var instantTargetId = ResolvePeerEnd(world, e.LastInstantCastTargetEnemyNetId, e.LastInstantCastTargetRole)?.GameObjectId;
-                enemy.Cast(e.LastInstantCastActionId, targetLocation: instantTargetLocation, castSeconds: 0f, targetId: instantTargetId);
-            }
-            // Edge-triggered like ModelState -- PlayAnimationTimeline/AttachLockonVfx
-            // are one-shot cues, so re-issuing them every snapshot even when
-            // unchanged would restart the same animation/VFX on a loop.
-            if (e.AnimationTimelineId is { } timelineId
-                && (!peerEnemyAnimationTimeline.TryGetValue(e.NetId, out var lastTimeline) || lastTimeline != timelineId))
-            {
-                peerEnemyAnimationTimeline[e.NetId] = timelineId;
-                DiagnosticLog.Info($"[Multiplayer] Peer: enemy NetId {e.NetId} (BNpcBase {e.BNpcBaseId}) AnimationTimelineId -> 0x{timelineId:X4}.");
-                enemy.PlayAnimationTimeline(timelineId);
-            }
-            if (e.NewLockonVfxIds.Count > 0)
-            {
-                DiagnosticLog.Info($"[Multiplayer] Peer: enemy NetId {e.NetId} (BNpcBase {e.BNpcBaseId}) NewLockonVfxIds -> [{string.Join(",", e.NewLockonVfxIds)}].");
-                foreach (var lockonId in e.NewLockonVfxIds)
-                    enemy.AttachLockonVfx(lockonId, persistent: false);
-            }
-        }
-        foreach (var staleId in peerEnemies.Keys.Where(id => !seenEnemyIds.Contains(id)).ToList())
-        {
-            DiagnosticLog.Info($"[Multiplayer] Peer: enemy NetId {staleId} no longer in snapshot -- despawning local doppel.");
-            peerEnemies[staleId].Despawn();
-            peerEnemies.Remove(staleId);
-            peerEnemyModelState.Remove(staleId);
-            peerEnemyLastLoggedStatuses.Remove(staleId);
-            peerEnemyAnimationTimeline.Remove(staleId);
-            peerEnemyLastInstantCastSeq.Remove(staleId);
-            peerEnemyLastCastSeq.Remove(staleId);
-        }
-
-        // UMAD P3 only (harmless no-op elsewhere -- no enemy will ever match
-        // BlackHole's BNpcBaseId in another scenario). A peer never runs the
-        // scenario's own Run_BlackHoleObstacles, so without this a debug-bot
-        // peer's own MoveTo pathing has no avoidance data and can cut straight
-        // through a black hole's damage radius mid-transit -- which the host
-        // would then apply DamageDown for, since the peer's own reported
-        // position is self-authoritative. Rebuilt from the already-synced
-        // peerEnemies rather than replaying the scenario's RNG state, so it
-        // can never drift from whatever the host is actually showing.
-        world.Obstacles.Clear();
-        // Diagnostic-only, paired with UmadP3BlackHoleScenario.Tick's own "Near black
-        // hole" log: that one is the host's belief (built off this peer's last
-        // self-reported pose, via SimNetworkPuppet.Position); this is the peer's own,
-        // true local position at the same real moment. If a future DamageDown dump
-        // shows the host's line but not this one at a comparable timestamp, the pose
-        // report was stale when it mattered -- if both show it, the peer's own
-        // pathing genuinely cut it close. localPlayer is null on the host (it drives
-        // its own bots directly, no self-pose loop), so this is peer-only already.
-        var localPlayer = Plugin.GameInstance.World.Party.Player;
-        foreach (var (netId, bh) in peerEnemies.Where(kvp => kvp.Value.BNpcBaseId == BNpcBaseId.BlackHole))
-        {
-            world.Obstacles.Add(new CircleObstacle(new Vector2(bh.Position.X, bh.Position.Z), UmadP3BlackHoleScenario.BlackHoleAvoidRadius));
-            if (localPlayer is null) continue;
-            var distSq = localPlayer.Placement().DistanceSq(bh.Position);
-            if (distSq < UmadP3BlackHoleScenario.NearBlackHoleLogRadius * UmadP3BlackHoleScenario.NearBlackHoleLogRadius)
-                DiagnosticLog.Info(
-                    $"[Multiplayer] Peer: local position ({localPlayer.Position.X:F2},{localPlayer.Position.Z:F2}) is {MathF.Sqrt(distSq):F2}y from black hole NetId {netId} at ({bh.Position.X:F2},{bh.Position.Z:F2}).");
-        }
-
-        // Debug-bot replay: Chaos/Exdeath might not have been replicated yet
-        // when TryStartDebugBotReplay first resolved them (WorldSnapshot and
-        // AiReplayStateMessage are independent flows) -- keep retrying here on
-        // every snapshot until each is found, cheap once it's no longer null.
-        if (debugShadowState is { } shadow)
-        {
-            shadow.ScenarioObjects.Chaos ??= peerEnemies.Values.FirstOrDefault(e => e.BNpcBaseId == BNpcBaseId.ChaosP3);
-            shadow.ScenarioObjects.Exdeath ??= peerEnemies.Values.FirstOrDefault(e => e.BNpcBaseId == BNpcBaseId.Exdeath);
-        }
-
-        var seenTetherIds = new HashSet<int>();
-        foreach (var t in snap.Tethers)
-        {
-            seenTetherIds.Add(t.NetId);
-            var a = ResolvePeerEnd(world, t.AEnemyNetId, t.ARole);
-            var b = ResolvePeerEnd(world, t.BEnemyNetId, t.BRole);
-            if (a == null && b == null) continue;
-            // Re-create on any endpoint change (e.g. a grabby tether's B going from
-            // unattached to a role once someone actually grabs it) -- SimTether's
-            // endpoints are fixed at construction, so there's no in-place update.
-            // Without this, a peer's local tether object -- and anything reading it,
-            // like UmadP3BlackHoleAi.PullTether's ReferenceEquals(t.B, player) check
-            // -- stays frozen at whatever endpoints were true the instant this NetId
-            // was first seen, for the rest of the run.
-            var aDesc = t.AEnemyNetId is { } aId ? $"enemy#{aId}" : t.ARole?.ToString() ?? "null";
-            var bDesc = t.BEnemyNetId is { } bId ? $"enemy#{bId}" : t.BRole?.ToString() ?? "null";
-            if (peerTethers.TryGetValue(t.NetId, out var existing))
-            {
-                if (ReferenceEquals(existing.A, a) && ReferenceEquals(existing.B, b)) continue;
-                DiagnosticLog.Info($"[Multiplayer] Peer: tether NetId {t.NetId} endpoint changed -- recreating (A={aDesc}, B={bDesc}).");
-                existing.Despawn();
-            }
-            else
-            {
-                DiagnosticLog.Info($"[Multiplayer] Peer: first snapshot of tether NetId {t.NetId} (TetherId {t.TetherId}) -- A={aDesc}, B={bDesc}.");
-            }
-            peerTethers[t.NetId] = world.Tether(a, b, t.TetherId);
-        }
-        foreach (var staleId in peerTethers.Keys.Where(id => !seenTetherIds.Contains(id)).ToList())
-        {
-            DiagnosticLog.Info($"[Multiplayer] Peer: tether NetId {staleId} no longer in snapshot -- despawning.");
-            peerTethers[staleId].Despawn();
-            peerTethers.Remove(staleId);
-        }
-
-        var myRole = MyClaimedRole;
-        foreach (var r in snap.Roles)
-        {
-            // Position is self-authoritative for our own claimed role (a real
-            // SimPlayer reports its own pose via SelfPoseMessage, never network-driven
-            // here) -- but statuses/lockon VFX are NOT: we never run any scenario
-            // logic ourselves, so nothing else would ever call AddStatus/
-            // AttachLockonVfx against our own real character. Reconcile those for
-            // every role, including our own.
-            if (r.Role != myRole && world.Party.Get(r.Role) is SimNetworkPuppet puppet)
-                puppet.ApplyNetworkPose(new Vector3(r.X, r.Y, r.Z), r.Rotation);
-
-            if (world.Party.Get(r.Role) is not { } member) continue;
-            var currentStatuses = member.ActiveStatusSnapshot;
-            if (!peerRoleReconciledStatusIds.TryGetValue(r.Role, out var reconciledIds))
-                peerRoleReconciledStatusIds[r.Role] = reconciledIds = new HashSet<ushort>();
-            foreach (var target in r.Statuses)
-            {
-                // Tracked regardless of whether AddStatus actually needs to run this
-                // tick, so the removal loop below recognizes it as host-managed even
-                // on a snapshot where nothing changed.
-                reconciledIds.Add(target.StatusId);
-                if (currentStatuses.Any(s => s.StatusId == target.StatusId && s.Stacks == target.Stacks)) continue;
-                member.AddStatus(target.StatusId, duration: target.RemainingTime, stacks: target.Stacks, overrideStacks: true);
-            }
-            // Only ever removes a statusId THIS reconciliation previously added --
-            // never diffs against the character's full ActiveStatusSnapshot the way
-            // the enemy path does, since a peer's own real character can carry
-            // statuses nothing here put there (see peerRoleReconciledStatusIds).
-            foreach (var trackedId in reconciledIds.ToList())
-            {
-                if (r.Statuses.Any(s => s.StatusId == trackedId)) continue;
-                member.RemoveStatus(trackedId);
-                reconciledIds.Remove(trackedId);
-            }
-            var statusKey = string.Join(",", r.Statuses.Select(s => $"{s.StatusId}:{s.Stacks}"));
-            if (!peerRoleLastLoggedStatuses.TryGetValue(r.Role, out var lastStatusKey) || lastStatusKey != statusKey)
-            {
-                peerRoleLastLoggedStatuses[r.Role] = statusKey;
-                DiagnosticLog.Info($"[Multiplayer] Peer: role {r.Role} statuses -> [{statusKey}].");
-            }
-            if (r.NewLockonVfxIds.Count > 0)
-            {
-                DiagnosticLog.Info($"[Multiplayer] Peer: role {r.Role} NewLockonVfxIds -> [{string.Join(",", r.NewLockonVfxIds)}].");
-                foreach (var lockonId in r.NewLockonVfxIds)
-                    member.AttachLockonVfx(lockonId, persistent: false);
-            }
-        }
-
-        var seenEventObjectIds = new HashSet<int>();
-        foreach (var o in snap.EventObjects)
-        {
-            seenEventObjectIds.Add(o.NetId);
-            if (!peerEventObjects.TryGetValue(o.NetId, out var eo))
-            {
-                var config = new EventObjectSpawnConfig
-                {
-                    EObjId = o.EObjId,
-                    Placement = new Placement(new Vector3(o.X, o.Y, o.Z), o.Rotation),
-                    TimelineState = o.TimelineState,
-                    SpawnVisible = true,
-                };
-                DiagnosticLog.Info($"[Multiplayer] Peer: first snapshot of event object NetId {o.NetId} -- EObj 0x{o.EObjId:X}, pos ({o.X:F2},{o.Y:F2},{o.Z:F2}), state {o.CurrentState} -- spawning local copy.");
-                eo = world.SpawnEventObject(config);
-                if (eo == null)
-                {
-                    DiagnosticLog.Warn($"[Multiplayer] Peer: SpawnEventObject returned null for NetId {o.NetId} (EObj 0x{o.EObjId:X}) -- skipping.");
-                    continue;
-                }
-                peerEventObjects[o.NetId] = eo;
-            }
-            eo.SetPosition(new Placement(new Vector3(o.X, o.Y, o.Z), o.Rotation));
-            // Edge-triggered like ModelState -- SetState is a plain field write with
-            // no rebuild to worry about, but re-issuing it every snapshot even
-            // when unchanged is pointless churn.
-            if (!peerEventObjectState.TryGetValue(o.NetId, out var lastState) || lastState != o.CurrentState)
-            {
-                peerEventObjectState[o.NetId] = o.CurrentState;
-                DiagnosticLog.Info($"[Multiplayer] Peer: event object NetId {o.NetId} (EObj 0x{o.EObjId:X}) CurrentState -> {o.CurrentState}.");
-                eo.SetState(o.CurrentState);
-            }
-        }
-        foreach (var staleId in peerEventObjects.Keys.Where(id => !seenEventObjectIds.Contains(id)).ToList())
-        {
-            DiagnosticLog.Info($"[Multiplayer] Peer: event object NetId {staleId} no longer in snapshot -- despawning local copy.");
-            peerEventObjects[staleId].Despawn();
-            peerEventObjects.Remove(staleId);
-            peerEventObjectState.Remove(staleId);
-        }
-    }
-
-    private SimCharacter? ResolvePeerEnd(SimWorld world, int? enemyNetId, PartyRole? role)
-    {
-        if (enemyNetId is { } id) return peerEnemies.GetValueOrDefault(id);
-        if (role is { } r) return world.Party.Get(r);
-        return null;
-    }
-
-    private void OnRoleKilledReceived(RoleKilledMessage msg)
-    {
-        if (IsHost) return;
-        DiagnosticLog.Info($"[Multiplayer] {msg.Role} killed: {msg.Cause}");
-        if (Plugin.GameInstance.World.Party.Get(msg.Role) is ISimPartyMember member)
-            Plugin.GameInstance.Kill(member, msg.Cause);
-        else
-            DiagnosticLog.Debug($"[Multiplayer] RoleKilled for {msg.Role} but that slot isn't an ISimPartyMember locally -- dropping.");
-    }
-
-    private void OnEndReceived(EndMessage msg)
-    {
-        if (IsHost) return;
-        // Not gated on `running`: OnWorldSnapshotReceived spawns/tracks enemies for
-        // any non-host peer regardless of role or running state (e.g. a spectator who
-        // joined but never claimed a role -- see the MyClaimedRole check in
-        // OnStartReceived, which leaves running false for them for the whole run).
-        // That peer still has live snapshot-spawned doppels/enmity-list entries to
-        // tear down here; bailing on !running left them stuck on screen for exactly
-        // that case. The IsInInstance check below already safely no-ops this for a
-        // peer who hasn't actually entered the zone yet.
-        DiagnosticLog.Info($"[Multiplayer] Peer received EndMessage (ReturnedToInn={msg.ReturnedToInn}).");
-        running = false;
-        StopDebugBotReplay();
-        // If our own deferred zone entry (RunScenarioAsPeer) hasn't actually
-        // completed yet, there's nothing to leave or reset -- and calling
-        // Leave() here specifically would be actively harmful: Unload()
-        // assumes a zone was entered (it restores the real character to the
-        // position ZoneSession.Enter() saved) and would instead teleport them
-        // to garbage coordinates. Reset() has no equivalent issue (peers never
-        // set ActiveScenario, so its own teleport-back-if-needed check never
-        // fires), but skip it too here for symmetry -- there's truly nothing
-        // to reset.
-        if (!Plugin.GameInstance.World.Map.IsInInstance)
-        {
-            DiagnosticLog.Info("[Multiplayer] EndMessage received before our own deferred zone entry completed -- nothing to leave/reset.");
-            return;
-        }
-        // Mirror whichever the host actually did -- Leave() if they left the
-        // zone entirely, or Reset() to match them staying in-zone (ready for a
-        // quick re-Start) instead of always hard-kicking to the inn.
-        if (msg.ReturnedToInn)
-            Plugin.GameInstance.Leave();
-        else
-            Plugin.GameInstance.Reset();
-    }
-
-    // ---- Debug: bot-controlled peer replay ----------------------------------
-
-    // Host-only, edge-triggered once per run. LastState isn't guaranteed set on
-    // the very first Tick() after StartScenario -- RunScenarioAsHost's actual
-    // work (including constructing the scenario's *State) is deferred via
-    // Plugin.Framework.Run, same reasoning as peerEnteredInstance below -- so
-    // this polls instead of reading it synchronously right after the call.
-    // Sent unconditionally: the host never knows or cares which peers, if any,
-    // are using it locally. One case per IScenario.SupportsMultiplayer scenario --
-    // add a new one here alongside a new *AiReplayStateMessage and
-    // *State.FromNetworkReplay when porting another scenario.
-    private void TrySendAiReplayState()
-    {
-        switch (Plugin.GameInstance.Scenarios[Session.ScenarioIndex])
-        {
-            case UmadP3BlackHoleScenario p3Scenario:
-                if (p3Scenario.LastState is not { } p3State) return;
-                aiReplayStateSent = true;
-                DiagnosticLog.Info("[Multiplayer] Host: broadcasting AiReplayState for this run.");
-                _ = relay!.SendAsync(new AiReplayStateMessage(
-                    p3State.Roles.List, p3State.StackTargets.List, p3State.SlapAttacks.ToArray(),
-                    p3State.KefkaPosition.Select(d => d.RadiansFromNorth).ToArray(), p3State.ImplosionAttack));
-                break;
-            case UmadP2ForsakenScenario p2Scenario:
-                if (p2Scenario.LastState is not { } p2State) return;
-                aiReplayStateSent = true;
-                DiagnosticLog.Info("[Multiplayer] Host: broadcasting AiReplayState for this run.");
-                _ = relay!.SendAsync(new P2AiReplayStateMessage(
-                    p2State.EndAttacks, p2State.NewNorth.RadiansFromNorth, p2State.Rotation, p2State.Lockons));
-                break;
-            case UmadP4KefkaSaysScenario p4Scenario:
-                if (p4Scenario.LastState is not { } p4State) return;
-                aiReplayStateSent = true;
-                DiagnosticLog.Info("[Multiplayer] Host: broadcasting AiReplayState for this run.");
-                _ = relay!.SendAsync(new P4AiReplayStateMessage(
-                    p4State.Mystery.Select(m => m.BlizzardOffset).ToArray(),
-                    p4State.Mystery.Select(m => m.LightningOffset).ToArray(),
-                    p4State.Mystery.Select(m => m.LightningOrientation).ToArray(),
-                    p4State.Wave1First, p4State.Wave1.List, p4State.Wave1True,
-                    p4State.Wave2.List, p4State.Wave2True,
-                    p4State.InfernoMystery.IsTrue, p4State.TsunamiMystery.IsTrue,
-                    p4State.Wave3.List, p4State.Wounds,
-                    p4State.Antilights[0].Antilight == Antilight.White,
-                    p4State.NeoExdeathDirection.RadiansFromNorth));
-                break;
-            case UmadP5ExaflaresScenario p5Scenario:
-                if (p5Scenario.LastState is not { } p5State) return;
-                aiReplayStateSent = true;
-                DiagnosticLog.Info("[Multiplayer] Host: broadcasting AiReplayState for this run.");
-                _ = relay!.SendAsync(new P5AiReplayStateMessage(p5State.LeftOrder.ToArray(), p5State.RightOrder.ToArray()));
-                break;
-        }
-    }
-
-    // Guards the AiStrats[Session.SelectedAi] indexing every multi-strat
-    // scenario's replay branch does below. StartScenario/MultiplayerWindow
-    // already refuse to broadcast an out-of-range SelectedAi (see
-    // MainWindow.HasStartableStrat), but a peer has no independent way to
-    // verify what the host sent, so this is the last line of defense against
-    // an IndexOutOfRangeException here.
-    private bool IsValidAiIndex(IScenario scenario) =>
-        Session.SelectedAi >= 0 && Session.SelectedAi < scenario.AiStrats.Count;
-
-    // Peer-only, idempotent, edge-triggered once per run: fires once both the
-    // host's *AiReplayStateMessage has arrived (pending*AiReplayState) and our
-    // own zone/party is actually ready (peerEnteredInstance) -- arrival order
-    // between those two isn't guaranteed, so both call sites (Dispatch and
-    // Tick) funnel through here. A no-op entirely unless debug-bot mode is on.
-    // Branches on the active scenario the same way TrySendAiReplayState does.
-    private void TryStartDebugBotReplay()
-    {
-        if (!debugBotControlled || debugBotReplayStarted) return;
-        if (!peerEnteredInstance) return;
-        if (MyClaimedRole is not { } myRole) return;
-        var world = Plugin.GameInstance.World;
-
-        switch (Plugin.GameInstance.Scenarios[Session.ScenarioIndex])
-        {
-            case UmadP3BlackHoleScenario when pendingAiReplayState is { } msg:
-            {
-                debugBotReplayStarted = true;
-                DiagnosticLog.Info($"[Multiplayer] Peer: starting debug-bot replay for {myRole}.");
-                var shadowState = UmadP3BlackHoleState.FromNetworkReplay(
-                    world, msg.Roles, msg.StackTargets, msg.SlapAttacks, msg.KefkaPositionRadians, msg.ImplosionAttack);
-                // Chaos/Exdeath might not have been replicated yet (WorldSnapshot
-                // and this message are independent flows) -- OnWorldSnapshotReceived
-                // keeps retrying this resolution against debugShadowState below as
-                // new enemies come in, so a still-null boss here isn't a lost
-                // cause, just not needed until each choreography step that reads
-                // it, tens of seconds into the fight at the earliest.
-                shadowState.ScenarioObjects.Chaos = peerEnemies.Values.FirstOrDefault(e => e.BNpcBaseId == BNpcBaseId.ChaosP3);
-                shadowState.ScenarioObjects.Exdeath = peerEnemies.Values.FirstOrDefault(e => e.BNpcBaseId == BNpcBaseId.Exdeath);
-                debugShadowState = shadowState;
-                DebugBotControl.Enabled = true;
-                new UmadP3BlackHoleAi().Run(shadowState, world);
-                break;
-            }
-            case UmadP2ForsakenScenario p2Scenario when pendingP2AiReplayState is { } p2Msg:
-            {
-                // Belt-and-suspenders: StartScenario/MultiplayerWindow already refuse
-                // to broadcast an out-of-range SelectedAi (see HasStartableStrat), but
-                // marking replay "started" either way (rather than just `break`) stops
-                // this from silently re-attempting -- and re-logging nothing -- every
-                // single frame if it's ever somehow reached anyway.
-                debugBotReplayStarted = true;
-                if (!IsValidAiIndex(p2Scenario))
-                {
-                    DiagnosticLog.Warn($"[Multiplayer] Peer: SelectedAi {Session.SelectedAi} is out of range for {p2Scenario.Name} ({p2Scenario.AiStrats.Count} strats) -- skipping debug-bot replay.");
-                    break;
-                }
-                DiagnosticLog.Info($"[Multiplayer] Peer: starting debug-bot replay for {myRole}.");
-                var shadowState = UmadP2ForsakenState.FromNetworkReplay(
-                    p2Msg.EndAttacks, p2Msg.NewNorthRadians, p2Msg.Rotation, p2Msg.Lockons);
-                debugShadowStateP2 = shadowState;
-                DebugBotControl.Enabled = true;
-                ((IScenarioAi<UmadP2ForsakenState>)p2Scenario.AiStrats[Session.SelectedAi]).Run(shadowState, world);
-                break;
-            }
-            case UmadP4KefkaSaysScenario p4Scenario when pendingP4AiReplayState is { } p4Msg:
-            {
-                debugBotReplayStarted = true;
-                if (!IsValidAiIndex(p4Scenario))
-                {
-                    DiagnosticLog.Warn($"[Multiplayer] Peer: SelectedAi {Session.SelectedAi} is out of range for {p4Scenario.Name} ({p4Scenario.AiStrats.Count} strats) -- skipping debug-bot replay.");
-                    break;
-                }
-                DiagnosticLog.Info($"[Multiplayer] Peer: starting debug-bot replay for {myRole}.");
-                var shadowState = UmadP4KefkaSaysState.FromNetworkReplay(
-                    world.Party, p4Msg.MysteryBlizzardOffset, p4Msg.MysteryLightningOffset, p4Msg.MysteryLightningOrientation,
-                    p4Msg.Wave1First, p4Msg.Wave1, p4Msg.Wave1True, p4Msg.Wave2, p4Msg.Wave2True,
-                    p4Msg.InfernoIsTrue, p4Msg.TsunamiIsTrue, p4Msg.Wave3, p4Msg.Wounds,
-                    p4Msg.Antilight0IsWhite, p4Msg.NeoExdeathDirectionRadians);
-                debugShadowStateP4 = shadowState;
-                DebugBotControl.Enabled = true;
-                ((IScenarioAi<UmadP4KefkaSaysState>)p4Scenario.AiStrats[Session.SelectedAi]).Run(shadowState, world);
-                break;
-            }
-            case UmadP5ExaflaresScenario p5Scenario when pendingP5AiReplayState is { } p5Msg:
-            {
-                debugBotReplayStarted = true;
-                if (!IsValidAiIndex(p5Scenario))
-                {
-                    DiagnosticLog.Warn($"[Multiplayer] Peer: SelectedAi {Session.SelectedAi} is out of range for {p5Scenario.Name} ({p5Scenario.AiStrats.Count} strats) -- skipping debug-bot replay.");
-                    break;
-                }
-                DiagnosticLog.Info($"[Multiplayer] Peer: starting debug-bot replay for {myRole}.");
-                // A fresh, peer-owned EventScheduler -- NOT the host's -- since
-                // UmadP5ExaflaresAi schedules its dodges directly onto it, and
-                // nothing would ever drive one shared with anything else. See
-                // MultiplayerManager.Tick's P5-specific branch just below for
-                // what actually ticks it every frame (mirrors UmadP5ExaflaresScenario.Tick,
-                // which normally does this but never runs on a peer).
-                var shadowState = UmadP5ExaflaresState.FromNetworkReplay(p5Msg.LeftOrder, p5Msg.RightOrder, new EventScheduler());
-                debugShadowStateP5 = shadowState;
-                DebugBotControl.Enabled = true;
-                ((IScenarioAi<UmadP5ExaflaresState>)p5Scenario.AiStrats[Session.SelectedAi]).Run(shadowState, world);
-                break;
-            }
-        }
-    }
-
-    // Clears just the current run's replay state, not the debugBotControlled
-    // toggle itself (a sticky lobby preference) -- called whenever running
-    // stops for any reason, so a debug-bot peer's real character always
-    // regains normal control the instant the fight ends rather than staying
-    // bot-driven while standing in an empty arena or back in the inn.
-    private void StopDebugBotReplay()
-    {
-        if (debugBotReplayStarted) DiagnosticLog.Info("[Multiplayer] Peer: stopping debug-bot replay.");
-        DebugBotControl.Enabled = false;
-        pendingAiReplayState = null;
-        pendingP2AiReplayState = null;
-        pendingP4AiReplayState = null;
-        pendingP5AiReplayState = null;
-        debugShadowState = null;
-        debugShadowStateP2 = null;
-        debugShadowStateP4 = null;
-        debugShadowStateP5 = null;
-        debugBotReplayStarted = false;
-    }
-
-    // ---- Message pump -------------------------------------------------------
-
-    // Ordering, not just marshalling onto the Framework thread, is why this exists.
-    // RelayClient.ReceiveLoopAsync invokes MessageReceived synchronously and
-    // sequentially, one message fully handled before the next ReceiveAsync -- so this
-    // enqueue always happens in true wire-arrival order. But firing a separate
-    // Plugin.Framework.Run(() => Dispatch(message)) per message (the old approach)
-    // does NOT preserve that order once queued: Dalamud's Framework.Run schedules
-    // onto ThreadBoundTaskScheduler, whose Run() iterates a ConcurrentDictionary of
-    // pending tasks -- not insertion order. Confirmed via AnoMech-DamageDebug dumps:
-    // a burst of MapEffectMessages sent host-side in order 1-8,0 (and by then
-    // already verified arriving wire-ordered peer-side, per RelayClient's own FIFO
-    // send queue) still executed out of order once each hit its own Framework.Run,
-    // scrambling a peer's replicated arena-color transition (P2 Forsaken's
-    // gold->black swap). Draining this queue once per Tick (already called
-    // unconditionally every frame from Plugin.OnFrameworkUpdate) sidesteps that
-    // scheduler entirely instead of fighting its ordering.
-    private readonly ConcurrentQueue<MpMessage> pendingMessages = new();
-
-    private void OnMessageReceivedOffThread(MpMessage message) => pendingMessages.Enqueue(message);
-
-    private void DrainPendingMessages()
-    {
-        while (pendingMessages.TryDequeue(out var message))
-            Dispatch(message);
-    }
-
-    // `source` is the specific RelayClient instance this event came from --
-    // compared against the current `relay` field so a stale event from a
-    // client we've already torn down (LeaveSession) or superseded (a
-    // newer reconnect attempt winning the race) is ignored rather than
-    // reprocessed or double-triggering another reconnect loop. This also
-    // doubles as the "was this intentional" check: LeaveSession always nulls
-    // `relay` before the corresponding Disconnected event can be dispatched
-    // (Dispose's cancellation unwinds on a background thread and this handler
-    // is itself marshalled to run on a later framework tick), so a manual
-    // Leave's own Disconnected(null) never reaches past this guard.
-    private void OnDisconnectedOffThread(RelayClient source, Exception? failure)
-        => Plugin.Framework.Run(() =>
-        {
-            if (!ReferenceEquals(relay, source)) return;
-            source.Dispose();
-            relay = null;
-            disconnectedSinceMs ??= Environment.TickCount64;
-            if (failure != null)
-            {
-                DiagnosticLog.Warn($"[Multiplayer] Disconnected: {failure.Message}");
-                ConnectionError = failure.Message;
-            }
-            else
-            {
-                DiagnosticLog.Info("[Multiplayer] Disconnected (no failure reported -- socket just closed).");
-            }
-            LobbyChanged?.Invoke();
-            BeginReconnect();
-        });
-
-    private void Dispatch(MpMessage message)
-    {
-        // A single malformed-but-parseable message (an unexpected null, a role
-        // enum out of range, whatever) must not take down the framework tick
-        // pump that every other plugin system also shares -- log and move on
-        // rather than letting one bad packet cascade.
-        try
-        {
-            DispatchCore(message);
-        }
-        catch (Exception e)
-        {
-            DiagnosticLog.Warn($"[Multiplayer] Error handling {message.GetType().Name}: {e}");
-        }
-    }
-
-    private void DispatchCore(MpMessage message)
-    {
-        // Every message type the host actually broadcasts to everyone (as
-        // opposed to a fellow peer's request that the relay's dumb fan-out
-        // happens to deliver to us too, e.g. another peer's ClaimRoleMessage --
-        // this switch just never matches those on a non-host client). Used to
-        // drive the host's own roster-row liveness (see lastHostMessageMs);
-        // must be kept in sync with the `when !IsHost` cases below.
-        // SessionEndedMessage deliberately excluded -- unlike everything else
-        // here it isn't guaranteed to have come from the host (any peer can
-        // send it). When it *is* the host leaving, it's moot anyway since
-        // receiving it tears the whole session down a few lines later
-        // regardless; when it's a departing peer instead, it plainly isn't a
-        // host message at all and must not be mistaken for one.
-        if (!IsHost && message is LobbyStateMessage or StartMessage or WorldSnapshotMessage
-            or RoleKilledMessage or EndMessage or PingMessage or PeerStatusMessage
-            or AiReplayStateMessage or P2AiReplayStateMessage or P4AiReplayStateMessage or P5AiReplayStateMessage)
-        {
-            lastHostMessageMs = Environment.TickCount64;
-            everHeardFromHost = true;
-        }
-
-        switch (message)
-        {
-            // Host-authoritative: only the host acts on requests other clients send.
-            case HelloMessage hello when IsHost:
-                peerLastSeenMs[hello.PeerId] = Environment.TickCount64;
-                Session.Names[hello.PeerId] = hello.DisplayName;
-                Session.Builds[hello.PeerId] = new PeerBuildInfo(hello.Version, hello.Checksum);
-                DiagnosticLog.Info($"[Multiplayer] Hello from {hello.PeerId} ({hello.DisplayName}), build {hello.Version} ({new PeerBuildInfo(hello.Version, hello.Checksum).ShortChecksum}), mismatch={IsVersionMismatched(hello.PeerId)}.");
-                BroadcastLobbyState();
-                break;
-            case ClaimRoleMessage claim when IsHost:
-                peerLastSeenMs[claim.PeerId] = Environment.TickCount64;
-                ApplyClaim(claim.PeerId, claim.Role);
-                break;
-            case ReleaseRoleMessage release when IsHost:
-                peerLastSeenMs[release.PeerId] = Environment.TickCount64;
-                ApplyRelease(release.PeerId);
-                break;
-            case SelfPoseMessage pose when IsHost:
-                peerLastSeenMs[pose.PeerId] = Environment.TickCount64;
-                OnSelfPoseReceived(pose);
-                break;
-            case PongMessage pong when IsHost:
-                peerLastSeenMs[pong.PeerId] = Environment.TickCount64;
-                peerLatencyMs[pong.PeerId] = Environment.TickCount64 - pong.SentAtMs;
-                break;
-
-            // Peer-facing broadcasts from the host.
-            case LobbyStateMessage lobby when !IsHost:
-                Session.ApplyLobbyState(lobby);
-                LobbyChanged?.Invoke();
-                // The host never re-sends Start to an already-open connection --
-                // only a fresh StartMessage triggers OnStartReceived. Without this,
-                // a peer who connects after Start already fired (a late join, or a
-                // manual rejoin after their connection dropped mid-fight) would sit
-                // forever on "Connected -- waiting for the host to start" even
-                // though the fight is already running. OnStartReceived is itself
-                // idempotent, so this is safe to also fall through for it on a
-                // normal fresh start (arrives just before StartMessage does).
-                if (lobby.Started && MyClaimedRole != null)
-                    OnStartReceived();
-                break;
-            case StartMessage when !IsHost:
-                OnStartReceived();
-                break;
-            case StartCheckMessage when !IsHost:
-            {
-                var reason = CheckOwnStartReadiness();
-                _ = relay?.SendAsync(new StartCheckResponseMessage(MyPeerId, reason == null, reason));
-                break;
-            }
-            case StartCheckResponseMessage resp when IsHost:
-                DiagnosticLog.Info($"[Multiplayer] StartCheck reply from {Session.NameOf(resp.PeerId)}: ready={resp.Ready}{(resp.Reason is { } r ? $" ({r})" : "")}.");
-                // Remove(...) returning false means either a duplicate/stale
-                // reply or one that arrived after the timeout already gave up
-                // on this peer -- either way there's nothing left to do with it.
-                if (pendingStartResponses == null || !pendingStartResponses.Remove(resp.PeerId)) break;
-                if (!resp.Ready) startCheckFailures[resp.PeerId] = resp.Reason ?? "not ready";
-                if (pendingStartResponses.Count == 0) FinishStartCheck();
-                break;
-            case WorldSnapshotMessage snap when !IsHost:
-                OnWorldSnapshotReceived(snap);
-                break;
-            case RoleKilledMessage killed when !IsHost:
-                OnRoleKilledReceived(killed);
-                break;
-            case EndMessage end when !IsHost:
-                OnEndReceived(end);
-                break;
-            case PingMessage ping when !IsHost:
-                _ = relay?.SendAsync(new PongMessage(MyPeerId, ping.SentAtMs));
-                break;
-            case PeerStatusMessage status when !IsHost:
-                peerStatuses.Clear();
-                foreach (var (id, entry) in status.Statuses)
-                    peerStatuses[id] = entry;
-                break;
-            // No `when !IsHost` guard -- when the HOST leaves it ends the
-            // session for the whole group (including any other peers), so
-            // that branch has to run regardless of the recipient's role. A
-            // departing peer, by contrast, only ever shrinks the roster (see
-            // RemovePeer) -- the rest of the group keeps going.
-            case SessionEndedMessage ended when ended.PeerId == Session.HostId:
-            {
-                // Read the sender's name before LeaveSessionInternal wipes Session out from under it.
-                var who = Session.NameOf(ended.PeerId);
-                DiagnosticLog.Info($"[Multiplayer] Host {who} left -- session ending for the whole group.");
-                // IsInInstance guard, not running -- same reasoning as
-                // LeaveRequestMessage below: a Reset earlier in the session (e.g. the
-                // peer clicking Reset before the host later leaves) clears running
-                // while leaving the group stuck in-instance, so gating on running here
-                // would skip Leave() and strand this peer in-instance with no session
-                // left to recover through. IsInInstance is only ever true once a zone
-                // was genuinely entered (MapController.TryLoad), which is the actual
-                // precondition Leave() -> Unload() needs (it restores the real
-                // character to the position ZoneSession.Enter() saved).
-                if (Plugin.GameInstance.World.Map.IsInInstance) Plugin.GameInstance.Leave();
-                LeaveSessionInternal(notifyOthers: false);
-                SessionEndReason = $"{who} left -- session ended.";
-                LobbyChanged?.Invoke();
-                break;
-            }
-            case SessionEndedMessage ended when IsHost:
-                RemovePeer(ended.PeerId);
-                break;
-            case ResetRequestMessage req when IsHost:
-                DiagnosticLog.Info($"[Multiplayer] {Session.NameOf(req.PeerId)} requested a reset.");
-                Plugin.GameInstance.Reset();
-                break;
-            // IsInInstance guard (not running -- Leave() must still work after
-            // a Reset, which clears running while leaving the group stuck
-            // in-instance with nothing to show for it and no way back short of
-            // disbanding the whole session): Leave() -> Unload() assumes a zone
-            // was actually entered (it restores the real character to the
-            // position ZoneSession.Enter() saved), and IsInInstance is only
-            // ever set true once that has genuinely happened (MapController.
-            // TryLoad), so it alone is the correct signal here.
-            case LeaveRequestMessage req when IsHost:
-                DiagnosticLog.Info($"[Multiplayer] {Session.NameOf(req.PeerId)} requested to leave the instance.");
-                if (Plugin.GameInstance.World.Map.IsInInstance)
-                    Plugin.GameInstance.Leave();
-                // Unconditional, even when IsInInstance was already false above (e.g.
-                // the host left via their own Leave button, possibly after an earlier
-                // Reset, before this request arrived) -- the old code silently dropped
-                // the request in exactly that case, leaving the requesting peer's own
-                // Leave button waiting forever for a response the host never sent.
-                // Always returnedToInn: true -- this handler's whole purpose is
-                // granting a leave request, regardless of whether Leave() above was
-                // just queued (its deferred World.Map.Unload() hasn't run yet, so
-                // reading IsInInstance here would still see the pre-Leave state).
-                BroadcastRunEnded(returnedToInn: true);
-                break;
-            case AiReplayStateMessage state when !IsHost:
-                pendingAiReplayState = state;
-                TryStartDebugBotReplay();
-                break;
-            case P2AiReplayStateMessage p2State when !IsHost:
-                pendingP2AiReplayState = p2State;
-                TryStartDebugBotReplay();
-                break;
-            // Re-syncs the one field in P2AiReplayStateMessage's snapshot that
-            // actually changes mid-fight -- see P2LockonsUpdateMessage's own doc
-            // comment. Dropped if the shadow state doesn't exist yet: that only
-            // happens if this update raced ahead of the replay actually starting,
-            // in which case the initial P2AiReplayStateMessage.Lockons snapshot
-            // (sent before any tower has had a chance to resolve and reassign
-            // anything) already carries the same value.
-            case P2LockonsUpdateMessage lockonsUpdate when !IsHost:
-                if (debugShadowStateP2 is { } p2Shadow)
-                    p2Shadow.Lockons = lockonsUpdate.Lockons;
-                break;
-            // Pure replays of the exact host-side call -- see
-            // MapEffectMessage/MapDirectorUpdateMessage's doc comment. The
-            // handlers subscribed in SubscribeMapEventsOnce gate on IsHost, so
-            // this can't loop back into a re-broadcast: AddEffect/DirectorUpdate
-            // fire the same MapController events on a peer's own local call too,
-            // but that peer's own handler is a no-op since IsHost is false there.
-            case MapEffectMessage effect when !IsHost:
-                DiagnosticLog.Info($"[Multiplayer] Peer: applying MapEffect packetFlags=0x{effect.PacketFlags:X8} index=0x{effect.Index:X}.");
-                Plugin.GameInstance.World.Map.AddEffect(effect.PacketFlags, effect.Index);
-                break;
-            case MapDirectorUpdateMessage directorUpdate when !IsHost:
-                DiagnosticLog.Info($"[Multiplayer] Peer: applying MapDirectorUpdate category=0x{directorUpdate.Category:X8}.");
-                Plugin.GameInstance.World.Map.DirectorUpdate(
-                    directorUpdate.Category, directorUpdate.Arg1, directorUpdate.Arg2,
-                    directorUpdate.Arg3, directorUpdate.Arg4, directorUpdate.Arg5, directorUpdate.Arg6);
-                break;
-            case SetWeatherMessage weather when !IsHost:
-                DiagnosticLog.Info($"[Multiplayer] Peer: applying SetWeather weatherId={weather.WeatherId} transition={weather.Transition}.");
-                Plugin.GameInstance.World.Map.SetWeather(weather.WeatherId, weather.Transition);
-                break;
-            case P4AiReplayStateMessage p4State when !IsHost:
-                pendingP4AiReplayState = p4State;
-                TryStartDebugBotReplay();
-                break;
-            case P5AiReplayStateMessage p5State when !IsHost:
-                pendingP5AiReplayState = p5State;
-                TryStartDebugBotReplay();
-                break;
-        }
-    }
 }
